sentence_transformers/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["Deo <liangzhanzhao@metrodata.cn>"]
description = "grpc server & client for sentence transformers"
name = "mdt-grpc-sentence-transformers"
readme = "readme.md"
version = "0.0.0"

[tool.poetry.dependencies]
betterproto = "^v2.0.0b6"
grpclib = {extras = ["protobuf"], version = "^0.4.5"}
langchain-core = ">=0.2.0"
langchain-community = ">=0.2.0"
# TODO(Deo): need cudnn9 for newer onnxruntime-gpu
onnxruntime-gpu = "<=1.18.0"
pydantic-settings = "^2.0.3"
python = ">=3.8.1,<=3.12"
sentence-transformers = ">=3.0.1"
optimum = {extras = ["onnxruntime-gpu"], version = "^1.20.0"}
huggingface-hub = ">=0.23.0"
torch = "2.3.1"
